refactor: fix model namings and module distribution
