+++ simba/core/models/spectrum_encoder.py
